+++ src/alphadb/utils/types.py
@@ -13,7 +13,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from typing import Literal
+from typing import Literal, List, Tuple
 
 DatabaseColumnTypeIntVariations = Literal["INT", "TINYINT", "BIGINT"]
 DatabaseColumnTypeTextVariations = Literal["TEXT", "LONGTEXT"]
@@ -26,3 +26,15 @@ DatabaseColumnType = Literal[
     "DATETIME",
     "JSON",
 ]
+
+Method = Literal["createtable"] | Literal["altertable"]
+Method.__doc__ = "Available version source methods"
+
+ValidationIssueLevel = Literal["LOW"] | Literal["HIGH"] | Literal["CRITICAL"]
+ValidationIssueLevel.__doc__ = """
+LOW: Will work, but will not have any effect on the database.
+HIGH: Will still work, but might produce a different result than desired.
+CRITICAL: Will not execute.
+"""
+ValidationIssuesList = List[Tuple[ValidationIssueLevel, str]]
+ValidationIssuesList.__doc__ = "An issue is structured like Tuple(level, issue)"